requirements.txt
Flask==2.1.0
google-cloud-datastore==2.7.1
google-cloud-bigquery==2.34.4